+++ weblate/trans/models/comment.py
@@ -57,6 +57,23 @@ class CommentManager(models.Manager):
             unit.translation.component.report_source_bugs
         )
 
+    def copy(self, project):
+        """Copy comments to new project
+
+        This is used on moving component to other project and ensures nothing
+        is lost. We don't actually look where the comment belongs as it
+        would make the operation really expensive and it should be done in the
+        cleanup cron job.
+        """
+        for comment in self.all():
+            Comment.objects.create(
+                project=project,
+                comment=comment.comment,
+                content_hash=comment.content_hash,
+                user=comment.user,
+                language=comment.language,
+            )
+
 
 @python_2_unicode_compatible
 class Comment(UnitData, UserDisplayMixin):